% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexsurv-helpers.R
\name{add.covs}
\alias{add.covs}
\title{Add covs}
\usage{
add.covs(x, pars, beta, X, transform = FALSE)
}
\arguments{
\item{x}{Object of type \code{flexsurvreg}.}

\item{pars}{Matrix of parameters}

\item{beta}{Matrix of betas}

\item{X}{Model-frame}

\item{transform}{Apply the inverse-transform associated with this family?}
}
\value{
Matrix of basepars
}
\description{
Helper for \code{predict.flexsurvreg}.
}

