% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysurv-package.R
\name{convert_time_cols_to_binned}
\alias{convert_time_cols_to_binned}
\title{Take a dataframe, and convert its time column(s) to binned versions.}
\usage{
convert_time_cols_to_binned(data, time_period, formula_lhs)
}
\arguments{
\item{data}{A data.frame}

\item{time_period}{Number for bin-unit}

\item{formula_lhs}{The expression on the left-hand-side of the formula, which generates the `Surv` object}
}
\value{
The data, with the relevant time-columns now binned.
}
\description{
Take a dataframe, and convert its time column(s) to binned versions.
}
