% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysurv-package.R
\name{convert_to_cr_survreg}
\alias{convert_to_cr_survreg}
\title{Create a `cr_survreg` model from a custom list of models}
\usage{
convert_to_cr_survreg(list_of_models, data, time_col_name, event_col_name,
  time_lb_col_name = NULL, time_dependent_config = list(time_start_col_name
  = NULL, id_col_name = NULL, time_dependent_col_names = NULL))
}
\arguments{
\item{list_of_models}{A list, whose names correspond to each possible type of event. These
should correspond to the values in the \code{event_col_name} column. Each element is a
survival-regression model that can predict survival probabilities.}

\item{data}{A data.frame}

\item{time_col_name}{A character indicating the column-name of the 'time' column.}

\item{event_col_name}{A character indicating the column-name for events. The values in this
column (either factor or character), should match the names of the names of
\code{list_of_models}, except for the value indicating censoring.}

\item{time_lb_col_name}{Optional. A character indicating the column-name of the time
'lower-bound'.}

\item{time_dependent_config}{If your data is in a 'counting process' format (i.e., there are
multiple rows per 'person', with a 'start' column specifying right-truncation), you should
supply this. A list with three entries: `time_start_col_name`, `id_col_name`, and
`time_dependent_col_names`.}
}
\value{
An object of type \code{cr_survreg}, with plot and summary methods.
}
\description{
Instead of using a predetermined type of survival-regression model for the submodels in
cr_survreg, you can use any survival-regression model with a valid `predict` method (one that
takes \code{type='survival'} and \code{times}; see \code{?predict.flexsurvreg}).
}
