% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexsurv-helpers.R
\name{form.model.matrix}
\alias{form.model.matrix}
\title{Form model-matrix}
\usage{
form.model.matrix(object, newdata)
}
\arguments{
\item{object}{Object of type \code{flexsurvreg}.}

\item{newdata}{A dataframe}
}
\value{
A model.matrix.
}
\description{
Helper for \code{predict.flexsurvreg}.
}
