% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysurv-package.R
\name{get_terms_mapper}
\alias{get_terms_mapper}
\title{Get a function which maps from column names to model-terms and vice versa}
\usage{
get_terms_mapper(formula, data)
}
\arguments{
\item{formula}{The formula to be passed to model.frame}

\item{data}{The data.frame}
}
\value{
A function that gives the model-term(s) for each original-column, or gives the
  original-column(s) for each model-term. You can also get original column-names from model-frame
  names (e.g., instead of specific levels of factor, just factor(variable) gets mapped to
  variable).
}
\description{
Get a function which maps from column names to model-terms and vice versa
}
