% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysurv-package.R
\name{get_terms_mappings}
\alias{get_terms_mappings}
\title{Get mapping from variables/column to formula terms}
\usage{
get_terms_mappings(terms, data)
}
\arguments{
\item{terms}{A 'terms' object}

\item{data}{A dataframe}
}
\value{
A tidy dataframe with mapping.
}
\description{
Get mapping from variables/column to formula terms
}
