% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysurv-package.R
\name{ggplot.tidysurv}
\alias{ggplot.tidysurv}
\title{Ggplot method for tidysurv objects}
\usage{
\method{ggplot}{tidysurv}(data, mapping = NULL, ...,
  environment = parent.frame())
}
\arguments{
\item{data}{A tidysurv object}

\item{mapping}{An aesthetic mapping for ggplot, created by one of the \code{aes} functions.}

\item{...}{Further arguments to be passed to ggplot}

\item{environment}{See \code{?ggplot2::ggplot}}
}
\value{
A ggplot object without any layers, but with sensible defaults for aesthetic mappings:
  each distinct predictor in a group (with distinct colors), and (if a competing risk model) each
  state in a separate group (with distinct line-types).
}
\description{
Ggplot method for tidysurv objects
}

