% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysurv-package.R
\name{interval_censor_helper}
\alias{interval_censor_helper}
\title{Create Surv object of type 'interval' given upper and lower bounds of event-times.}
\usage{
interval_censor_helper(time, time_lb, event)
}
\arguments{
\item{time}{The time-column}

\item{time_lb}{The lower-bound for interval censoring.}

\item{event}{Event indicator (numeric or logical only)}
}
\value{
A \code{Surv} object with interval-censoring.
}
\description{
Create Surv object of type 'interval' given upper and lower bounds of event-times.
}

