% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysurv-package.R
\name{merge_formulae}
\alias{merge_formulae}
\title{Merge formulae}
\usage{
merge_formulae(forms, data)
}
\arguments{
\item{forms}{List of formulae}

\item{data}{Data.frame}
}
\value{
Merged formula
}
\description{
Merge formulae
}
