% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysurv-package.R
\name{plot.tidysurv}
\alias{plot.tidysurv}
\title{Plot a tidysurv object}
\usage{
\method{plot}{tidysurv}(x, mapping = NULL, type = "cumulative",
  states = NULL, ...)
}
\arguments{
\item{x}{A tidysurv object}

\item{mapping}{An aesthetic mapping for ggplot, created by one of the \code{aes} functions.}

\item{type}{What type of plot? Cumulative/survival, or rate? If the latter, need to have set
\code{time_period} in \code{tidysurv}.}

\item{states}{For a competing risks model: which states should be included in the graph?}

\item{...}{Further arguments to be passed to ggplot}
}
\value{
A ggplot object, with lines for each survival curve, and ribbons for confidence bands.
  Defaults to a separate survival curve (with a distinct color) for each unique combination of
  predictors that was passed to \code{tidysurv}. Also grouped by `.surv_state` for competing risk models,
  with the linetype indicating which state.
}
\description{
Plot a tidysurv object
}
