% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexsurv-helpers.R
\name{predict.flexsurvreg}
\alias{predict.flexsurvreg}
\title{Predict method for `flexsurvreg` models}
\usage{
\method{predict}{flexsurvreg}(object, newdata, times, type = "survival",
  start = NULL, ...)
}
\arguments{
\item{object}{Object of type \code{flexsurvreg}.}

\item{newdata}{A dataframe to apply predictions to.}

\item{type}{What type of prediction? Options are "survival", "cumhaz", and "hazard".}

\item{start}{Optional. A numeric vector of start (truncation) times.}

\item{...}{Ignored.}

\item{t}{A numeric vector of times for which to return predictions, with length equal to
\code{nrow(newdata)}}
}
\value{
A numeric vector of predictions
}
\description{
This function returns predictions from a \code{flexsurvreg} object. This can be a
convenient/faster alternative to \code{summary.flexsurvreg}, when the number of observations with
unique combinations of covariates is large.
}
