% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysurv-package.R
\name{tidy.survfit}
\alias{tidy.survfit}
\title{Tidy an object of class `survfit`}
\usage{
\method{tidy}{survfit}(x, ...)
}
\arguments{
\item{x}{An object of class `survfit`}

\item{...}{Ignored.}
}
\value{
A tidy dataframe
}
\description{
This function is/should be found in the `broom` package, but occasionally I want an immediate bug-
fix, and so I place it here.
}
