% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysurv-package.R
\name{tidysurv.cr_survreg}
\alias{tidysurv.cr_survreg}
\title{Create a tidy-surv object for an object of class \code{cr_survreg}}
\usage{
\method{tidysurv}{cr_survreg}(object, newdata = NULL, group_vars = NULL,
  time_period = NULL, id_col_name = NULL, time_dependent_vars = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{cr_survreg}}

\item{newdata}{Optional. A dataframe.}

\item{group_vars}{A character-vector specifying column(s) to group-by when creating the
survival-curves.}

\item{...}{Ignored.}

\item{other_vars_fun}{All covariates not in \code{group_vars} will have this function applied to
them. Should be a function that takes a numeric vector and returns a numeric vector with a single unique value.
Defaults to `mean` with na-removal.}
}
\value{
An object of class \code{cr_survreg}
}
\description{
Create a tidy-surv object for an object of class \code{cr_survreg}
}
