% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysurv-package.R
\name{tidysurv.formula}
\alias{tidysurv.formula}
\title{Make a tidy survfit object from a formula}
\usage{
\method{tidysurv}{formula}(object, data, na.action = na.exclude,
  time_period = NULL, max_num_levels = 200, ...)
}
\arguments{
\item{object}{A formula, as would be specified for \code{survival::survfit.formula}}

\item{data}{A dataframe}

\item{na.action}{a missing-data filter function}

\item{time_period}{You can choose to bin events into time-periods. This can be helpful if you want
plot the rate within a time-bin, rather than the cumulative rate of incidence.}

\item{max_num_levels}{Since this function creates a separate survival-curve for each distinct
level of predictors (the right hand side of the \code{formula}), it can sometimes be the case
that you'll accidentally create hundreds or thousands of distinct survival curves, because you
accidentally passed the wrong column-name to the formula. This argument ensures that you don't
accidentally eat up time computing these curves when you didnt mean to. Default is 200.}

\item{...}{Further arguments to be passed to \code{survival::survfit.formula}}
}
\value{
A tidy dataframe
}
\description{
Make a tidy survfit object from a formula
}
