% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysurv-package.R
\docType{package}
\name{tidysurv}
\alias{tidysurv}
\alias{tidysurv-package}
\title{tidysurv package}
\usage{
tidysurv(object, ...)
}
\arguments{
\item{object}{Current methods support a formula or an object of type \code{cr_survreg}}

\item{...}{Further arguments to be passed to methods.}
}
\value{
A tidy dataframe, describing survival over time
}
\description{
tidysurv package

See \code{?tidysurv.formula} and \code{tidysurv.cr_survreg} for more details.
}

