% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysurv-package.R
\name{update_aes}
\alias{update_aes}
\title{Update ggplot aes}
\usage{
update_aes(old_aes, new_aes)
}
\arguments{
\item{old_aes}{The old aes}

\item{new_aes}{New aes to merge in. Replaces any conflicts in old_aes}
}
\value{
An updated aes for ggplot mapping argument.
}
\description{
Update ggplot aes
}

