% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysurv-package.R
\name{tidysurv.cr_survreg}
\alias{tidysurv.cr_survreg}
\title{Create a tidy-surv object for an object of class \code{cr_survreg}}
\usage{
\method{tidysurv}{cr_survreg}(object, newdata = NULL, group_vars = NULL,
  time_period = NULL, id_col_name = NULL, time_dependent_vars = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{cr_survreg}}

\item{newdata}{Optional. A dataframe.}

\item{group_vars}{A character-vector specifying column(s) to group-by when creating the
survival-curves.}

\item{time_period}{This bins the time-values into groups of this width, which is needed for
plotting rates and time-dependent variables.}

\item{id_col_name}{Character-string of id-column. Only needed if originally supplied value
doesn't work for `newdata`}

\item{time_dependent_vars}{Character-strings for time-dependent covariates. Instead of taking the
overall mean, takes the mean at each time-point.}

\item{...}{Ignored.}
}
\value{
An object of class \code{cr_survreg}
}
\description{
Create a tidy-surv object for an object of class \code{cr_survreg}
}
